ship: iteration 1 — resolve all quality findings
