+++ .agents/skills/quality/SKILL.md
@@ -119,11 +119,13 @@ are handled by the synthesis step below, not a separate phase.
    Fix correctness findings and Track B judo moves alike. If a fix is genuinely
    large (a multi-file extraction, a schema migration), it is still yours to do —
    do it here, in this run, not "as a follow-up".
-7. **Re-review until clean.** If step 6 changed code, re-run Track A on the *new*
-   diff. New accepted findings → verify (4), apply (6), re-check. Stop when a
-   pass yields no new accepted findings. A re-review count is never a reason to
-   defer a verified finding or move it to the gate. This catches fix-induced
-   regressions before `/test` or `/ship`.
+7. **Re-review until clean.** If step 6 changed code, re-run **both mandatory
+   tracks, A and B,** on the *new* diff. New accepted findings → verify (4),
+   apply (6), re-check with both tracks again. Stop only when the same pass
+   yields no new accepted findings from either track. A re-review count is never
+   a reason to defer a verified finding or move it to the gate. This catches
+   fix-induced correctness regressions and maintainability debt before `/test`
+   or `/ship`.
 8. **Gate — the narrow exception, not the escape hatch.** Only two kinds of
    accepted finding may go to the gate unfixed:
    - it needs a **product decision you cannot make** (which of two valid
@@ -166,15 +168,18 @@ is not behavior-preserving on a branch that was not asked to change behavior.
 Empty is the expected outcome. "Structural / large / out of scope" is not a
 gate reason — those get fixed above.
 
-When empty, print `- Empty.` and omit the table. Never leave an example or
-placeholder row that another skill could mistake for a live gate.
+When empty, print exactly:
 
-| Severity | file:line | Finding | Which gate reason | Decision needed |
-|---|---|---|---|---|
-| Blocker | ... | ... | product decision | which behavior do you want? |
+- Empty.
 
-Next: /test (turn each Blocker/High into a named regression test).
+Do not print a table. When non-empty, replace `- Empty.` with a table containing
+only real findings and these columns: Severity, file:line, Finding, Which gate
+reason, Decision needed. Never leave an example or placeholder row that another
+skill could mistake for a live gate.
+
+Next: /test (itemize every accepted correctness finding and give each a named
+regression test or explicit alternate verification).
 
 **Before you print this:** every accepted finding is either in "Auto-applied" or
-in this table. If one is in neither, go back to step 6 and fix it.
+the Gate section. If one is in neither, go back to step 6 and fix it.
 ```
