+++ apps/desktop/src/main/services/prs/prService.ts
@@ -56,8 +56,6 @@ import type {
   PrCommit,
   PrConflictAnalysis,
   PrCreationStrategy,
-  PrDetachedLane,
-  PrMergedBy,
   PrEventPayload,
   PrGroupMemberRole,
   PrHealth,
@@ -161,6 +159,14 @@ import { spawn } from "node:child_process";
 import { runGit, runGitMergeTree, runGitOrThrow } from "../git/git";
 import { shouldAttemptAdminMergeForRestError } from "./resolverUtils";
 import { deletePullRequestRowsByIds } from "./pullRequestRowCleanup";
+import {
+  deriveGithubSnapshotLaneLink,
+  deriveGithubSnapshotMergeFacts,
+  normalizeCount,
+  normalizeMergeMethod,
+  rowDetachedLane,
+  rowMergedBy,
+} from "./prRowMetadata";
 import { createGithubStackStore } from "./githubStackStore";
 import { extractFirstJsonObject } from "../ai/utils";
 import { buildIntegrationPreflight } from "./integrationPlanning";
@@ -1065,87 +1071,6 @@ function rowToSummary(row: PullRequestRow): PrSummary {
     mergeMethod: normalizeMergeMethod(row.merge_method),
     commitCount: normalizeCount(row.commit_count),
     changedFiles: normalizeCount(row.changed_files)
-  };
-}
-
-function normalizeCount(value: unknown): number | null {
-  if (value == null) return null;
-  const parsed = Number(value);
-  return Number.isFinite(parsed) && parsed >= 0 ? parsed : null;
-}
-
-function normalizeMergeMethod(value: unknown): MergeMethod | null {
-  return value === "squash" || value === "merge" || value === "rebase" ? value : null;
-}
-
-/**
- * Resolve the lane columns of a list row.
- *
- * A detached row reports NO live lane (`linkedLaneId`/`linkedLaneName` stay null) and
- * carries `detached` instead, so the renderer shows it as history rather than as a
- * mapping it could act on. A missing lane row yields a null name, never the raw lane
- * UUID — a bare id in the list is worse than no chip at all.
- */
-function deriveGithubSnapshotLaneLink(
-  linked: PullRequestRow | null,
-  laneById: Map<string, LaneSummary> | undefined,
-): Pick<GitHubPrListItem, "linkedLaneId" | "linkedLaneName" | "detached"> {
-  if (!linked) return { linkedLaneId: null, linkedLaneName: null, detached: null };
-  const detached = rowDetachedLane(linked);
-  if (detached) return { linkedLaneId: null, linkedLaneName: null, detached };
-  return {
-    linkedLaneId: linked.lane_id,
-    linkedLaneName: laneById?.get(linked.lane_id)?.name ?? null,
-    detached: null,
-  };
-}
-
-/** Merge outcome + size, so a merged row can describe itself without a GitHub call. */
-function deriveGithubSnapshotMergeFacts(
-  linked: PullRequestRow | null,
-): Pick<
-  GitHubPrListItem,
-  "mergedAt" | "mergedBy" | "mergeMethod" | "additions" | "deletions" | "commitCount" | "changedFiles"
-> {
-  return {
-    mergedAt: linked?.merged_at ?? null,
-    mergedBy: linked ? rowMergedBy(linked) : null,
-    mergeMethod: normalizeMergeMethod(linked?.merge_method),
-    additions: normalizeCount(linked?.additions),
-    deletions: normalizeCount(linked?.deletions),
-    commitCount: normalizeCount(linked?.commit_count),
-    changedFiles: normalizeCount(linked?.changed_files),
-  };
-}
-
-function rowMergedBy(row: PullRequestRow): PrMergedBy | null {
-  const login = String(row.merged_by_login ?? "").trim();
-  if (!login) return null;
-  return { login, avatarUrl: row.merged_by_avatar_url ?? null };
-}
-
-/**
- * Rehydrate the lane provenance frozen at detach time. Returns null for live rows.
- * A malformed or missing provenance blob still yields a usable record — the lane name
- * is the part the UI leads with, and zeroed counts are simply not rendered.
- */
-function rowDetachedLane(row: PullRequestRow): PrDetachedLane | null {
-  const at = String(row.detached_at ?? "").trim();
-  if (!at) return null;
-  let counts: { chats?: unknown; artifacts?: unknown; checkpoints?: unknown } = {};
-  try {
-    const parsed = JSON.parse(String(row.detached_provenance ?? "{}"));
-    if (parsed && typeof parsed === "object") counts = parsed as typeof counts;
-  } catch {
-    /* a corrupt blob must not hide the lane name */
-  }
-  return {
-    at,
-    laneName: row.detached_lane_name ?? null,
-    laneColor: row.detached_lane_color ?? null,
-    chats: normalizeCount(counts.chats) ?? 0,
-    artifacts: normalizeCount(counts.artifacts) ?? 0,
-    checkpoints: normalizeCount(counts.checkpoints) ?? 0,
   };
 }
